workspace/agent-dev-team/requirements.txt
# AG2 (AutoGen) 核心依赖
ag2>=0.4.0

# AI模型接口
openai>=1.0.0
anthropic>=0.21.0

# 工具和实用库
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0

# 代码执行和安全
docker>=7.0.0

# 日志和监控
rich>=13.0.0
loguru>=0.7.0

# 测试
pytest>=8.0.0
pytest-asyncio>=0.23.0
